got rid of repeated available mission code
